4-空间数据分析/其他案例/旅行规划问题/pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eviltransform>=0.1.1",
    "folium>=0.20.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
